backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
pymongo==4.6.0
mongoengine==0.27.0
python-decouple==3.8
django-cors-headers==4.2.0
Pillow==10.0.0
python-dateutil==2.8.2
requests==2.31.0
